Add initial import from Druntime
The code imported here is from DMD commit 19899e9ce38910690ed9a4f15b1714f50d51f94e.
It is current as of 2023-10-03, as core.stdcpp hasn't been changed in two months.
The changes made to that code are as follow:
- Renamed modules and import from core.stdcpp.x to stdcpp.x;
- Added a github workflow loosely based on runnable_cxx.yml (in dmd);
- Added a dub.json file to server as build system;
- Added a stdcpp.test package to server as package for unittests;
- Added a single vector test which does not compile on POSIX;

This import is made to allow Emmanuel, a SAOC student, to start
working on improving C++ interop by providing easy to use bindings
for the STL types. Note that examples and documentation are also
planned, especially regarding choosing what runtime to use.