Add vector.cpp and link to it in unittest
We need an explicit instantiation of vector to link against,
so that C++ symbols are available. It's an extra step,
but any sizeable project will have this requirement one
way or another, and it is the most foolproof approach
(the other approach is to reproduce what std::vector
does in its C++ code while managing the 3 pointers).

For the moment, just add an extras/ folder and assume
unittests will need to add an explicit instantiation
when using symbols. In the future we can make that code
more generic, or even better, auto-generated.